nvidia_mlserver_test_copy/requirements.txt
mlserver
# fastapi
# uvicorn
# python-multipart
# numpy
soundfile
nemo_toolkit[asr]
resampy
huggingface-hub==0.23.2